Refactoring and trying to get test which touch a test databast to run sequentially instead of parrelel
Refactored method RunQuery into RunTransactionQuery and made the same method use a transaction instead of a raw query

Added the line [Collection("Sequential")] in front of every test class that needs to be run sequentially instead of parrallel. This is Xunit annotations but i know if it really worked.
In this context i also added the CollectionDefinition.cs class in hopes of defining this collection. Otherwise Xunit might think that each Collection named Sequential might be a new collection and therefore not work as intended